Incorporation de DTO et changement des int en Long

+++ src/main/java/ca/cal/tp1/modele/CD.java
@@ -13,6 +13,6 @@ import lombok.NoArgsConstructor;
 @DiscriminatorValue("CD")
 public class CD extends Document {
     private String artiste;
-    private int duree;
+    private Long duree;
     private String genre;
 }
